Fix int32 pointer-arithmetic overflow in ssd_bmm.py
Same overflow class as the rest of this series (see
https://github.com/state-spaces/mamba/issues/1015): Triton specializes
each pointer-arithmetic operand as 32-bit based on its own value at
specialization time, so a chained multiplication that should be 64-bit
wraps once the true offset exceeds 2**31 - 1.

Cast the relevant program-id-derived index to tl.int64 immediately after
tl.program_id() in both kernels in ssd_bmm.py, covering the batch-axis
term that an earlier pass had missed in each. Cheap, unconditional, no
measurable regression on the common contiguous/short-sequence case.

Regression tests cover _bmm_chunk_fwd's batch-axis overflow at genuine
overflow-triggering scale, comparing kernel output against a closed-form
known answer. Verified to actually catch the regression by temporarily
reverting the relevant int64 cast and confirming the test fails before
restoring the fix.

+++ mamba_ssm/ops/triton/ssd_bmm.py
@@ -49,8 +49,9 @@ def _bmm_chunk_fwd_kernel(
     HAS_SEQ_IDX: tl.constexpr,
     BLOCK_SIZE_M: tl.constexpr, BLOCK_SIZE_N: tl.constexpr, BLOCK_SIZE_K: tl.constexpr,
 ):
-    pid_b = tl.program_id(axis=1)
-    pid_ch = tl.program_id(axis=2)
+    # int64 to avoid int32 overflow, see https://github.com/state-spaces/mamba/issues/1015
+    pid_b = tl.program_id(axis=1).to(tl.int64)
+    pid_ch = tl.program_id(axis=2).to(tl.int64)
     pid_c = pid_ch // ngroups
     pid_h = pid_ch - pid_c * ngroups
     num_pid_n = tl.cdiv(chunk_size, BLOCK_SIZE_N)
@@ -122,8 +123,9 @@ def _bmm_chunk_bwd_kernel(
     HAS_RESIDUAL: tl.constexpr,
     BLOCK_SIZE_M: tl.constexpr, BLOCK_SIZE_N: tl.constexpr, BLOCK_SIZE_CS: tl.constexpr,
 ):
-    pid_b = tl.program_id(axis=1)
-    pid_ch = tl.program_id(axis=2)
+    # int64 to avoid int32 overflow, see https://github.com/state-spaces/mamba/issues/1015
+    pid_b = tl.program_id(axis=1).to(tl.int64)
+    pid_ch = tl.program_id(axis=2).to(tl.int64)
     pid_c = pid_ch // ngroups
     pid_h = pid_ch - pid_c * ngroups
     num_pid_n = tl.cdiv(K, BLOCK_SIZE_N)
